rewrote board implementation for lights out

+++ src/v2/framework/Puzzle.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface Puzzle {
     List<? extends Action<? extends Puzzle>> getActions();
-    int heuristicCost();
+    long heuristicCost();
     boolean isGoal();
 }
